editoast: move waypoint to editoast_schemas
