+++ editoast/editoast_schemas/src/infra.rs
@@ -1,9 +1,10 @@
 mod direction;
 mod track_offset;
+mod waypoint;
 
 pub use direction::Direction;
-
 pub use track_offset::TrackOffset;
+pub use waypoint::Waypoint;
 
 editoast_common::schemas! {
     track_offset::schemas(),
