+++ editoast/src/converters/generate_routes.rs
@@ -9,6 +9,7 @@ use std::collections::HashMap;
 use crate::schema::*;
 use editoast_common::Identifier;
 use editoast_schemas::infra::Direction;
+use editoast_schemas::infra::Waypoint;
 use editoast_schemas::primitives::OSRDIdentified;
 
 /* Part 1: type definitions */
