+++ editoast/src/schema/mod.rs
@@ -21,7 +21,6 @@ pub use detector::Detector;
 pub use detector::DetectorCache;
 use editoast_schemas::infra::Direction;
 use editoast_schemas::primitives::OSRDIdentified;
-use editoast_schemas::primitives::OSRDObject;
 use editoast_schemas::primitives::ObjectType;
 pub use electrification::Electrification;
 pub use errors::InfraError;
@@ -78,65 +77,6 @@ editoast_common::schemas! {
     utils::schemas(),
     editoast_schemas::schemas(),
     operation::schemas(),
-}
-
-#[derive(Deserialize, Serialize, Clone, Debug, PartialEq, Eq, Hash)]
-#[serde(tag = "type", deny_unknown_fields)]
-pub enum Waypoint {
-    BufferStop { id: Identifier },
-    Detector { id: Identifier },
-}
-
-impl Waypoint {
-    /// Create a new detector stop waypoint
-    pub fn new_detector<T: AsRef<str>>(detector: T) -> Self {
-        Self::Detector {
-            id: detector.as_ref().into(),
-        }
-    }
-
-    /// Create a new buffer stop waypoint
-    pub fn new_buffer_stop<T: AsRef<str>>(bf: T) -> Self {
-        Self::BufferStop {
-            id: bf.as_ref().into(),
-        }
-    }
-
-    /// Return whether the waypoint is a detector
-    pub fn is_detector(&self) -> bool {
-        matches!(self, Waypoint::Detector { .. })
-    }
-
-    // Return whether the waypoint is a buffer stop
-    pub fn is_buffer_stop(&self) -> bool {
-        matches!(self, Waypoint::BufferStop { .. })
-    }
-}
-
-impl Default for Waypoint {
-    fn default() -> Self {
-        Self::Detector {
-            id: "InvalidRef".into(),
-        }
-    }
-}
-
-impl OSRDIdentified for Waypoint {
-    fn get_id(&self) -> &String {
-        match self {
-            Waypoint::BufferStop { id } => id,
-            Waypoint::Detector { id } => id,
-        }
-    }
-}
-
-impl OSRDObject for Waypoint {
-    fn get_type(&self) -> ObjectType {
-        match self {
-            Waypoint::BufferStop { .. } => ObjectType::BufferStop,
-            Waypoint::Detector { .. } => ObjectType::Detector,
-        }
-    }
 }
 
 #[derive(Debug, Derivative, Clone, Deserialize, Serialize, PartialEq, ToSchema)]
