+++ editoast/src/schema/route.rs
@@ -9,13 +9,13 @@ use super::Endpoint;
 use super::OSRDIdentified;
 use super::ObjectType;
 use super::TrackEndpoint;
-use super::Waypoint;
 use crate::infra_cache::Cache;
 use crate::infra_cache::Graph;
 use crate::infra_cache::InfraCache;
 use crate::infra_cache::ObjectCache;
 use editoast_common::Identifier;
 use editoast_schemas::infra::Direction;
+use editoast_schemas::infra::Waypoint;
 use editoast_schemas::primitives::OSRDTyped;
 
 #[derive(Debug, Derivative, Clone, Deserialize, Serialize, PartialEq, Eq)]
